requirements.txt
awsebcli==3.15.2
bokeh==1.0.4
botocore==1.12.171
cement==2.8.2
certifi==2019.3.9
chardet==3.0.4
colorama==0.3.9
cycler==0.10.0
Django==2.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.11
docutils==0.14
future==0.16.0
idna==2.7
Jinja2==2.10.1
jmespath==0.9.4
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.0
mysqlclient==1.4.2.post1
numpy==1.16.4
packaging==19.0
pandas==0.24.2
pathspec==0.5.9
Pillow==6.0.0
pyparsing==2.4.0
pypiwin32==223
python-dateutil==2.8.0
pytz==2019.1
pywin32==224
PyYAML==3.13
requests==2.20.1
scipy==1.3.0
seaborn==0.9.0
semantic-version==2.5.0
six==1.11.0
termcolor==1.1.0
tornado==6.0.2
urllib3==1.24.3
wcwidth==0.1.7
wincertstore==0.2
